packages/models-library/requirements/ci.txt
# Shortcut to install all packages for the contigous integration (CI) of 'models-library'
#
# - As ci.txt but w/ tests
#
# Usage:
#   pip install -r requirements/ci.txt
#

# installs base + tests requirements
--requirement _base.txt
--requirement _test.txt

# installs this repo's packages
simcore-postgres-database[migration] @ ../postgres-database/
pytest-simcore @ ../pytest-simcore

# current module
simcore-models-library @ .
